change double to decimal all amount related data type.

+++ ExpenseProcessingSystem/ExpenseProcessingSystem/Models/ExpenseEntryCashBreakdownModel.cs
@@ -13,7 +13,7 @@ namespace ExpenseProcessingSystem.Models
         public int CashBreak_ID { get; set; }
         public double CashBreak_Denomination { get; set; }
         public int CashBreak_NoPcs { get; set; }
-        public double CashBreak_Amount { get; set; }
+        public decimal CashBreak_Amount { get; set; }
         public ExpenseEntryDetailModel ExpenseEntryDetailModel { get; set; }
     }
 }
